put create-grid into its own file
