+++ lib/commons/dom/find-nearby-elms.js
@@ -1,4 +1,4 @@
-import { createGrid } from './get-rect-stack';
+import createGrid from './create-grid';
 
 function findNearbyElms(vNode, margin = 0) {
   /*eslint no-bitwise: 0*/
