+++ lib/commons/dom/get-element-stack.js
@@ -1,5 +1,6 @@
-import { createGrid, getRectStack } from './get-rect-stack';
+import { getRectStack } from './get-rect-stack';
 import { getNodeFromTree } from '../../core/utils';
+import createGrid from './create-grid';
 
 /**
  * Return all elements that are at the center bounding rect of the passed in node.
