+++ lib/commons/dom/get-text-element-stack.js
@@ -1,5 +1,6 @@
 import getElementStack from './get-element-stack';
-import { createGrid, getRectStack } from './get-rect-stack';
+import { getRectStack } from './get-rect-stack';
+import createGrid from './create-grid';
 import sanitize from '../text/sanitize';
 import { getNodeFromTree } from '../../core/utils';
