+++ lib/commons/dom/index.js
@@ -41,4 +41,4 @@ export { default as urlPropsFromAttribute } from './url-props-from-attribute';
 export { default as visuallyContains } from './visually-contains';
 export { default as visuallyOverlaps } from './visually-overlaps';
 export { default as visuallySort } from './visually-sort';
-export { createGrid } from './get-rect-stack';
+export { default as createGrid } from './create-grid';
