+++ lib/commons/dom/visually-sort.js
@@ -1,3 +1,4 @@
+import createGrid from './create-grid';
 /**
  * Visually sort nodes based on their stack order
  * References:
@@ -7,6 +8,7 @@
  */
 export default function visuallySort(a, b) {
   /*eslint no-bitwise: 0 */
+  createGrid(); // Because we need ._stackingOrder
   const length = Math.max(a._stackingOrder.length, b._stackingOrder.length);
 
   for (let i = 0; i < length; i++) {
